Beginning work on manViewHandler, got default and policy done. Also working on pnpm integration to see if its faster.
