+++ frontend/package.json
@@ -3,6 +3,12 @@
 	"version": "0.1.0",
 	"private": true,
 	"pnpm": {
+		"peerDependencyRules": {
+			"allowedVersions": {
+				"react": "17",
+				"react-dom": "17"
+			}
+		},
 		"overrides": {
 			"bezier-js": "4.0.3",
 			"immer": "9.0.6",
@@ -22,7 +28,7 @@
 	},
 	"dependencies": {
 		"@babel/core": "7.17.10",
-		"@datapunt/matomo-tracker-react": "0.3.1",
+		"@datapunt/matomo-tracker-react": "0.5.1",
 		"@date-io/date-fns": "1.3.13",
 		"@dod-advana/advana-jira-feedback": "0.4.15",
 		"@dod-advana/advana-platform-ui": "1.4.4",
