+++ frontend/src/components/modules/eda/edaMainViewUtilityHandler.js
@@ -1,0 +1,7 @@
+import { setState } from '../../../utils/sharedFunctions';
+
+const EDAMainViewUtilityHandler = {
+	handleCategoryTabChange: (props) => {},
+};
+
+export default EDAMainViewUtilityHandler;
