+++ frontend/src/components/modules/globalSearch/globalSearchMainViewUtilityHandler.js
@@ -1,0 +1,7 @@
+import { setState } from '../../../utils/sharedFunctions';
+
+const GlobalSearchMainViewUtilityHandler = {
+	handleCategoryTabChange: (props) => {},
+};
+
+export default GlobalSearchMainViewUtilityHandler;
