+++ frontend/src/components/modules/jbook/jbookMainViewUtilityHandler.js
@@ -1,0 +1,7 @@
+import { setState } from '../../../utils/sharedFunctions';
+
+const JBookMainViewUtilityHandler = {
+	handleCategoryTabChange: (props) => {},
+};
+
+export default JBookMainViewUtilityHandler;
