+++ frontend/src/components/modules/policy/policyMainViewHandler.js
@@ -1,14 +1,20 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import moment from 'moment';
 import GameChangerSearchMatrix from '../../searchMetrics/GCSearchMatrix';
 import GameChangerSideBar from '../../searchMetrics/GCSideBar';
 import DefaultGraphView from '../../graph/defaultGraphView';
-import defaultMainViewHandler from '../default/defaultMainViewHandler';
+import {
+	getMainView as defaultGetMainView,
+	handleDidYouMeanClicked,
+	styles,
+	handlePageLoad as defaultHandlePageLoad,
+	getViewNames as defaultGetViewNames,
+} from '../default/defaultMainViewHandler';
 import PolicyDocumentExplorer from './policyDocumentExplorer';
 import ViewHeader from '../../mainView/ViewHeader';
 import { trackEvent } from '../../telemetry/Matomo';
 import { Typography } from '@material-ui/core';
-import { setState, handleSaveFavoriteTopic } from '../../../utils/sharedFunctions';
+import { setState, handleSaveFavoriteTopic, getNonMainPageOuterContainer } from '../../../utils/sharedFunctions';
 import Permissions from '@dod-advana/advana-platform-ui/dist/utilities/permissions';
 import SearchSection from '../globalSearch/SearchSection';
 import LoadingIndicator from '@dod-advana/advana-platform-ui/dist/loading/LoadingIndicator';
@@ -21,6 +27,7 @@ import GetQAResults from '../default/qaResults';
 import {
 	getQueryVariable,
 	getTrackingNameForFactory,
+	PAGE_DISPLAYED,
 	RESULTS_PER_PAGE,
 	StyledCenterContainer,
 } from '../../../utils/gamechangerUtils';
@@ -38,76 +45,34 @@ import '../../mainView/main-view.css';
 import DefaultSeal from '../../mainView/img/GC Default Seal.png';
 import DefaultPub from '../../mainView/img/default_cov.png';
 import GamechangerUserManagementAPI from '../../api/GamechangerUserManagement';
-import GCAboutUs from '../../aboutUs/GCAboutUs';
-//import { last } from 'underscore';
+import SearchHandlerFactory from '../../factories/searchHandlerFactory';
+import UserProfileHandlerFactory from '../../factories/userProfileFactory';
+import LoadableVisibility from 'react-loadable-visibility/react-loadable';
 const _ = require('lodash');
 
 const gameChangerAPI = new GameChangerAPI();
 const gcUserManagementAPI = new GamechangerUserManagementAPI();
 
-const fullWidthCentered = {
-	width: '100%',
-	display: 'flex',
-	flexDirection: 'column',
-	justifyContent: 'center',
-	alignItems: 'center',
-};
+const AnalystTools = LoadableVisibility({
+	loader: () => import('../../analystTools'),
+	loading: () => {
+		return <LoadingIndicator customColor={gcOrange} />;
+	},
+});
 
-const styles = {
-	listViewBtn: {
-		minWidth: 0,
-		margin: '20px 0px 0px',
-		marginLeft: 10,
-		padding: '0px 7px 0',
-		fontSize: 20,
-		height: 34,
+const GCDataStatusTracker = LoadableVisibility({
+	loader: () => import('../../dataTracker/GCDataStatusTracker'),
+	loading: () => {
+		return <LoadingIndicator customColor={gcOrange} />;
 	},
-	cachedResultIcon: {
-		display: 'flex',
-		justifyContent: 'center',
-		padding: '0 0 1% 0',
-	},
-	searchResults: fullWidthCentered,
-	paginationWrapper: fullWidthCentered,
-	resultsCount: {
-		fontFamily: 'Noto Sans',
-		fontSize: 22,
-		fontWeight: 'bold',
-		color: '#131E43',
-		paddingTop: '10px',
-	},
-	subtext: {
-		color: '#8091A5',
-		fontSize: 12,
-	},
-	containerText: {
-		fontSize: 16,
-		fontWeight: 'bold',
-	},
-	checkboxPill: {
-		textAlign: 'center',
-		borderRadius: '10px',
-		paddingLeft: '10px',
-		paddingRight: '10px',
-		lineHeight: 1.2,
-		fontSize: '12px',
-		marginLeft: '10px',
-		border: '2px solid #bdccde',
-		backgroundColor: 'white',
-		boxSizing: 'border-box',
-		color: 'black',
-		minHeight: '35px',
-		display: 'flex',
-		alignItems: 'center',
-		justifyContent: 'center',
-		cursor: 'pointer',
-	},
-};
+});
 
-const handleDidYouMeanClicked = (didYouMean, state, dispatch) => {
-	trackEvent(getTrackingNameForFactory(state.cloneData.clone_name), 'SuggestionSelected', 'DidYouMean');
-	setState(dispatch, { searchText: didYouMean, runSearch: true });
-};
+const GCAboutUs = LoadableVisibility({
+	loader: () => import('../../aboutUs/GCAboutUs'),
+	loading: () => {
+		return <LoadingIndicator customColor={gcOrange} />;
+	},
+});
 
 const getSearchResults = (searchResultData, state, dispatch) => {
 	return _.map(searchResultData, (item, idx) => {
@@ -216,6 +181,7 @@ const handlePopPubs = async (pop_pubs, pop_pubs_inactive, state, dispatch, cance
 		setState(dispatch, { searchMajorPubs: filteredPubs });
 	}
 };
+
 const handleLastOpened = async (last_opened_docs, state, dispatch, cancelToken) => {
 	let cleanedDocs = [];
 	let filteredPubs = [];
@@ -261,6 +227,7 @@ const handleLastOpened = async (last_opened_docs, state, dispatch, cancelToken) 
 		setState(dispatch, { lastOpened: filteredPubs });
 	}
 };
+
 const handleRecDocs = async (rec_docs, state, dispatch, cancelToken) => {
 	let filteredPubs = [];
 	try {
@@ -357,780 +324,828 @@ const formatString = (text) => {
 	return _.truncate(titleCase, { length: 60, separator: /,?\.* +/ });
 };
 
-const PolicyMainViewHandler = {
-	async handlePageLoad(props) {
-		const { state, dispatch } = props;
-		await defaultMainViewHandler.handlePageLoad(props);
-		setState(dispatch, { loadingrecDocs: true });
-		setState(dispatch, { loadingLastOpened: true });
+const handlePageLoad = async (props) => {
+	const { state, dispatch } = props;
+	await defaultHandlePageLoad(props);
+	setState(dispatch, { loadingrecDocs: true });
+	setState(dispatch, { loadingLastOpened: true });
 
-		let topics = [];
-		let pop_pubs = [];
-		let pop_pubs_inactive = [];
-		let rec_docs = [];
+	let topics = [];
+	let pop_pubs = [];
+	let pop_pubs_inactive = [];
+	let rec_docs = [];
 
-		const user = await gcUserManagementAPI.getUserData();
-		const { favorite_documents = [], export_history = [], pdf_opened = [] } = user.data;
+	const user = await gcUserManagementAPI.getUserData();
+	const { favorite_documents = [], export_history = [], pdf_opened = [] } = user.data;
 
-		try {
-			const { data } = await gameChangerAPI.getHomepageEditorData({
-				favorite_documents,
-				export_history,
-				pdf_opened,
+	try {
+		const { data } = await gameChangerAPI.getHomepageEditorData({
+			favorite_documents,
+			export_history,
+			pdf_opened,
+		});
+		data.forEach((obj) => {
+			if (obj.key === 'homepage_topics') {
+				topics = JSON.parse(obj.value);
+			} else if (obj.key === 'homepage_popular_docs_inactive') {
+				pop_pubs_inactive = JSON.parse(obj.value);
+			} else if (obj.key === 'popular_docs') {
+				pop_pubs = obj.value;
+			} else if (obj.key === 'rec_docs') {
+				rec_docs = obj.value;
+			}
+		});
+	} catch (e) {
+		// Do nothing
+		console.log(e);
+	}
+	setState(
+		dispatch,
+		getQueryVariable('view', window.location.hash.toString()) === 'graph'
+			? { adminTopics: topics, currentViewName: 'Graph', runGraphSearch: true }
+			: { adminTopics: topics }
+	);
+	// handlePubs(pubs, state, dispatch);
+	handleSources(state, dispatch, props.cancelToken);
+	handlePopPubs(pop_pubs, pop_pubs_inactive, state, dispatch, props.cancelToken);
+	handleRecDocs(rec_docs, state, dispatch, props.cancelToken);
+	handleLastOpened(pdf_opened, state, dispatch, props.cancelToken);
+};
+
+const getMainView = (props) => {
+	return defaultGetMainView(props);
+};
+
+const renderHideTabs = (props) => {
+	const { state, dispatch, searchHandler } = props;
+	const {
+		adminTopics,
+		searchMajorPubs,
+		recDocs,
+		loadingrecDocs,
+		cloneData,
+		crawlerSources,
+		prevSearchText,
+		resetSettingsSwitch,
+		didYouMean,
+		loading,
+		userData,
+		recentSearches,
+		trending,
+		lastOpened = [],
+		loadingLastOpened = true,
+	} = state;
+
+	const showDidYouMean = didYouMean && !loading;
+	if (prevSearchText) {
+		if (!resetSettingsSwitch) {
+			dispatch({ type: 'RESET_SEARCH_SETTINGS' });
+			setState(dispatch, {
+				resetSettingsSwitch: true,
+				showSnackbar: true,
+				snackBarMsg: 'Search settings reset',
 			});
-			data.forEach((obj) => {
-				if (obj.key === 'homepage_topics') {
-					topics = JSON.parse(obj.value);
-				} else if (obj.key === 'homepage_popular_docs_inactive') {
-					pop_pubs_inactive = JSON.parse(obj.value);
-				} else if (obj.key === 'popular_docs') {
-					pop_pubs = obj.value;
-				} else if (obj.key === 'rec_docs') {
-					rec_docs = obj.value;
-				}
-			});
-		} catch (e) {
-			// Do nothing
-			console.log(e);
+			if (searchHandler) searchHandler.setSearchURL(state);
 		}
-		setState(
-			dispatch,
-			getQueryVariable('view', window.location.hash.toString()) === 'graph'
-				? { adminTopics: topics, currentViewName: 'Graph', runGraphSearch: true }
-				: { adminTopics: topics }
-		);
-		// handlePubs(pubs, state, dispatch);
-		handleSources(state, dispatch, props.cancelToken);
-		handlePopPubs(pop_pubs, pop_pubs_inactive, state, dispatch, props.cancelToken);
-		handleRecDocs(rec_docs, state, dispatch, props.cancelToken);
-		handleLastOpened(pdf_opened, state, dispatch, props.cancelToken);
-	},
+	}
 
-	getMainView(props) {
-		return defaultMainViewHandler.getMainView(props);
-	},
+	const { favorite_topics = [], favorite_searches = [] } = userData;
 
-	handleCategoryTabChange(props) {
-		defaultMainViewHandler.handleCategoryTabChange(props);
-	},
+	// const agencyPublications = ['Department of the United States Army', 'Department of the United States Navy', 'Department of the United States Marine Corp', 'Department of United States Air Force']
 
-	renderHideTabs(props) {
-		const { state, dispatch, searchHandler } = props;
-		const {
-			adminTopics,
-			searchMajorPubs,
-			recDocs,
-			loadingrecDocs,
-			cloneData,
-			crawlerSources,
-			prevSearchText,
-			resetSettingsSwitch,
-			didYouMean,
-			loading,
-			userData,
-			recentSearches,
-			trending,
-			lastOpened = [],
-			loadingLastOpened = true,
-		} = state;
+	let trendingLinks = [];
 
-		const showDidYouMean = didYouMean && !loading;
-		if (prevSearchText) {
-			if (!resetSettingsSwitch) {
-				dispatch({ type: 'RESET_SEARCH_SETTINGS' });
-				setState(dispatch, {
-					resetSettingsSwitch: true,
-					showSnackbar: true,
-					snackBarMsg: 'Search settings reset',
+	if (trending) {
+		trending.data.forEach((search) => {
+			if (search.search) {
+				trendingLinks.push({
+					search: search.search.replaceAll('&#039;', '"'),
+					count: search.count,
+					favorite: false,
 				});
-				if (searchHandler) searchHandler.setSearchURL(state);
-			}
-		}
-
-		const { favorite_topics = [], favorite_searches = [] } = userData;
-
-		// const agencyPublications = ['Department of the United States Army', 'Department of the United States Navy', 'Department of the United States Marine Corp', 'Department of United States Air Force']
-
-		let trendingLinks = [];
-
-		if (trending) {
-			trending.data.forEach((search) => {
-				if (search.search) {
-					trendingLinks.push({
-						search: search.search.replaceAll('&#039;', '"'),
-						count: search.count,
-						favorite: false,
-					});
-				}
-			});
-		}
-
-		trendingLinks.forEach(({ search }, idx) => {
-			favorite_searches.forEach((fav) => {
-				if (fav.search_text === search) {
-					trendingLinks[idx].favorite = true;
-				}
-			});
-		});
-
-		recentSearches.forEach((search, idx) => {
-			favorite_searches.forEach((fav) => {
-				recentSearches[idx].favorite = fav.tiny_url === search.tiny_url;
-			});
-		});
-
-		const favTopicNames = favorite_topics.map((item) => item.topic_name.toLowerCase());
-		adminTopics.forEach((topic, idx) => {
-			if (
-				_.find(favTopicNames, (item) => {
-					return item === topic.name.toLowerCase();
-				})
-			) {
-				adminTopics[idx].favorite = true;
-			} else {
-				adminTopics[idx].favorite = false;
 			}
 		});
+	}
 
-		return (
-			<div style={{ marginTop: '40px' }}>
-				{prevSearchText && (
-					<div style={{ margin: '10px auto', width: '67%' }}>
-						<div style={styles.resultsCount}>
-							<p style={{ fontWeight: 'normal', display: 'inline' }}>
-								Looks like we don't have any matches for{' '}
-							</p>
-							"{prevSearchText}"
-						</div>
+	trendingLinks.forEach(({ search }, idx) => {
+		favorite_searches.forEach((fav) => {
+			if (fav.search_text === search) {
+				trendingLinks[idx].favorite = true;
+			}
+		});
+	});
+
+	recentSearches.forEach((search, idx) => {
+		favorite_searches.forEach((fav) => {
+			recentSearches[idx].favorite = fav.tiny_url === search.tiny_url;
+		});
+	});
+
+	const favTopicNames = favorite_topics.map((item) => item.topic_name.toLowerCase());
+	adminTopics.forEach((topic, idx) => {
+		if (
+			_.find(favTopicNames, (item) => {
+				return item === topic.name.toLowerCase();
+			})
+		) {
+			adminTopics[idx].favorite = true;
+		} else {
+			adminTopics[idx].favorite = false;
+		}
+	});
+
+	return (
+		<div style={{ marginTop: '40px' }}>
+			{prevSearchText && (
+				<div style={{ margin: '10px auto', width: '67%' }}>
+					<div style={styles.resultsCount}>
+						<p style={{ fontWeight: 'normal', display: 'inline' }}>
+							Looks like we don't have any matches for{' '}
+						</p>
+						"{prevSearchText}"
 					</div>
-				)}
-				{showDidYouMean && (
-					<div
-						style={{
-							margin: '10px auto',
-							fontSize: '25px',
-							width: '67%',
-							paddingLeft: 'auto',
-						}}
-					>
-						Did you mean{' '}
-						<DidYouMean onClick={() => handleDidYouMeanClicked(didYouMean, state, dispatch)}>
-							{didYouMean}
-						</DidYouMean>
-						?
-					</div>
-				)}
-				<div style={{ margin: '0 70px 0 70px' }}>
-					<GameChangerThumbnailRow links={trendingLinks} title={'Trending Searches'} width={'300px'}>
-						{trendingLinks.map(({ search, favorite, count }, idx) => (
-							<TrendingSearchContainer
-								onClick={() => setState(dispatch, { searchText: search, runSearch: true })}
+				</div>
+			)}
+			{showDidYouMean && (
+				<div
+					style={{
+						margin: '10px auto',
+						fontSize: '25px',
+						width: '67%',
+						paddingLeft: 'auto',
+					}}
+				>
+					Did you mean{' '}
+					<DidYouMean onClick={() => handleDidYouMeanClicked(didYouMean, state, dispatch)}>
+						{didYouMean}
+					</DidYouMean>
+					?
+				</div>
+			)}
+			<div style={{ margin: '0 70px 0 70px' }}>
+				<GameChangerThumbnailRow links={trendingLinks} title={'Trending Searches'} width={'300px'}>
+					{trendingLinks.map(({ search, favorite, count }, idx) => (
+						<TrendingSearchContainer
+							onClick={() => setState(dispatch, { searchText: search, runSearch: true })}
+						>
+							<div
+								style={{
+									display: 'flex',
+									justifyContent: 'space-between',
+									wordSpacing: 3,
+									letterSpacing: 2,
+									marginTop: 8,
+								}}
 							>
+								<Typography style={styles.containerText}>
+									<i className="fa fa-search" style={{ marginRight: 12 }} />
+									{`   ${idx + 1}. ${search.length < 18 ? search : search.substring(0, 20) + '...'}`}
+								</Typography>
+							</div>
+						</TrendingSearchContainer>
+					))}
+				</GameChangerThumbnailRow>
+				<GameChangerThumbnailRow
+					links={adminTopics}
+					title={"Editor's Choice: Top Topics"}
+					width="100px"
+					style={{ marginLeft: '0' }}
+				>
+					{adminTopics.map((item, idx) => (
+						<div
+							style={styles.checkboxPill}
+							onClick={() => {
+								trackEvent(getTrackingNameForFactory(cloneData.clone_name), 'TopicOpened', item.name);
+								window.open(
+									`#/gamechanger-details?cloneName=${
+										cloneData.clone_name
+									}&type=topic&topicName=${item.name.toLowerCase()}`
+								);
+							}}
+						>
+							{item.name}
+							<i
+								className={item.favorite ? 'fa fa-star' : 'fa fa-star-o'}
+								style={{
+									color: item.favorite ? '#E9691D' : 'rgb(224,224,224)',
+									marginLeft: 10,
+									cursor: 'pointer',
+									fontSize: 20,
+								}}
+								onClick={(event) => {
+									event.stopPropagation();
+									handleSaveFavoriteTopic(item.name.toLowerCase(), '', !item.favorite, dispatch);
+								}}
+							/>
+						</div>
+					))}
+				</GameChangerThumbnailRow>
+				<GameChangerThumbnailRow links={recDocs} title="Recommended For You" width="215px">
+					{recDocs.length > 0 &&
+						recDocs[0].imgSrc &&
+						recDocs.map((pub) => (
+							<div className="topPublication">
+								{pub.imgSrc !== 'error' ? (
+									<img className="image" src={pub.imgSrc} alt="thumbnail" title={pub.name} />
+								) : (
+									<div className="image">{pub.name}</div>
+								)}
+
 								<div
-									style={{
-										display: 'flex',
-										justifyContent: 'space-between',
-										wordSpacing: 3,
-										letterSpacing: 2,
-										marginTop: 8,
+									className="hover-overlay"
+									onClick={() => {
+										trackEvent(
+											getTrackingNameForFactory(cloneData.clone_name),
+											'PublicationOpened',
+											pub.name
+										);
+										pub.imgSrc !== DefaultPub
+											? window.open(
+													`#/gamechanger-details?cloneName=${cloneData.clone_name}&type=document&documentName=${pub.id}`
+											  )
+											: setState(dispatch, { searchText: pub.name, runSearch: true });
 									}}
 								>
-									<Typography style={styles.containerText}>
-										<i className="fa fa-search" style={{ marginRight: 12 }} />
-										{`   ${idx + 1}. ${
-											search.length < 18 ? search : search.substring(0, 20) + '...'
-										}`}
-									</Typography>
+									<div className="hover-text">{formatString(pub.name)}</div>
 								</div>
-							</TrendingSearchContainer>
+							</div>
 						))}
-					</GameChangerThumbnailRow>
-					<GameChangerThumbnailRow
-						links={adminTopics}
-						title={"Editor's Choice: Top Topics"}
-						width="100px"
-						style={{ marginLeft: '0' }}
-					>
-						{adminTopics.map((item, idx) => (
-							<div
-								style={styles.checkboxPill}
+					{loadingrecDocs && recDocs.length === 0 && (
+						<div className="col-xs-12">
+							<LoadingIndicator
+								customColor={gcOrange}
+								inline={true}
+								containerStyle={{
+									height: '300px',
+									textAlign: 'center',
+									paddingTop: '75px',
+									paddingBottom: '75px',
+								}}
+							/>
+						</div>
+					)}
+					{!loadingrecDocs && recDocs.length === 0 && (
+						<div className="col-xs-12" style={{ height: '140px' }}>
+							<Typography style={styles.containerText}>
+								Try favoriting more documents to see your personalized recommendations.
+							</Typography>
+						</div>
+					)}
+				</GameChangerThumbnailRow>
+				<GameChangerThumbnailRow links={lastOpened} title="Recently Viewed" width="215px">
+					{lastOpened.length > 0 &&
+						lastOpened[0].imgSrc &&
+						lastOpened.map((pub) => (
+							<div className="topPublication">
+								{pub.imgSrc !== 'error' ? (
+									<img className="image" src={pub.imgSrc} alt="thumbnail" title={pub.name} />
+								) : (
+									<div className="image">{pub.name}</div>
+								)}
+
+								<div
+									className="hover-overlay"
+									onClick={() => {
+										trackEvent(
+											getTrackingNameForFactory(cloneData.clone_name),
+											'PublicationOpened',
+											pub.name
+										);
+										pub.imgSrc !== DefaultPub
+											? window.open(
+													`#/gamechanger-details?cloneName=${cloneData.clone_name}&type=document&documentName=${pub.id}`
+											  )
+											: setState(dispatch, { searchText: pub.name, runSearch: true });
+									}}
+								>
+									<div className="hover-text">{formatString(pub.name)}</div>
+								</div>
+							</div>
+						))}
+					{loadingLastOpened && lastOpened.length === 0 && (
+						<div className="col-xs-12">
+							<LoadingIndicator
+								customColor={gcOrange}
+								inline={true}
+								containerStyle={{
+									height: '300px',
+									textAlign: 'center',
+									paddingTop: '75px',
+									paddingBottom: '75px',
+								}}
+							/>
+						</div>
+					)}
+					{!loadingLastOpened && lastOpened.length === 0 && (
+						<div className="col-xs-12" style={{ height: '140px' }}>
+							<Typography style={styles.containerText}>No recent documents to show.</Typography>
+						</div>
+					)}
+				</GameChangerThumbnailRow>
+
+				<GameChangerThumbnailRow links={crawlerSources} title="Sources" width="300px">
+					{crawlerSources.length > 0 &&
+						crawlerSources[0].imgSrc &&
+						crawlerSources.map((source) => (
+							<SourceContainer
 								onClick={() => {
 									trackEvent(
 										getTrackingNameForFactory(cloneData.clone_name),
-										'TopicOpened',
-										item.name
+										'SourceOpened',
+										source.data_source_s
 									);
 									window.open(
 										`#/gamechanger-details?cloneName=${
 											cloneData.clone_name
-										}&type=topic&topicName=${item.name.toLowerCase()}`
+										}&type=source&sourceName=${source.data_source_s.toLowerCase()}`
 									);
 								}}
 							>
-								{item.name}
-								<i
-									className={item.favorite ? 'fa fa-star' : 'fa fa-star-o'}
+								<img src={source.imgSrc} alt={'crawler seal'}></img>
+								<Typography
 									style={{
-										color: item.favorite ? '#E9691D' : 'rgb(224,224,224)',
-										marginLeft: 10,
-										cursor: 'pointer',
-										fontSize: 20,
+										...styles.containerText,
+										color: '#313541',
+										alignSelf: 'center',
+										marginLeft: '20px',
 									}}
-									onClick={(event) => {
-										event.stopPropagation();
-										handleSaveFavoriteTopic(item.name.toLowerCase(), '', !item.favorite, dispatch);
-									}}
-								/>
-							</div>
-						))}
-					</GameChangerThumbnailRow>
-					<GameChangerThumbnailRow links={recDocs} title="Recommended For You" width="215px">
-						{recDocs.length > 0 &&
-							recDocs[0].imgSrc &&
-							recDocs.map((pub) => (
-								<div className="topPublication">
-									{pub.imgSrc !== 'error' ? (
-										<img className="image" src={pub.imgSrc} alt="thumbnail" title={pub.name} />
-									) : (
-										<div className="image">{pub.name}</div>
-									)}
-
-									<div
-										className="hover-overlay"
-										onClick={() => {
-											trackEvent(
-												getTrackingNameForFactory(cloneData.clone_name),
-												'PublicationOpened',
-												pub.name
-											);
-											pub.imgSrc !== DefaultPub
-												? window.open(
-														`#/gamechanger-details?cloneName=${cloneData.clone_name}&type=document&documentName=${pub.id}`
-												  )
-												: setState(dispatch, { searchText: pub.name, runSearch: true });
-										}}
-									>
-										<div className="hover-text">{formatString(pub.name)}</div>
-									</div>
-								</div>
-							))}
-						{loadingrecDocs && recDocs.length === 0 && (
-							<div className="col-xs-12">
-								<LoadingIndicator
-									customColor={gcOrange}
-									inline={true}
-									containerStyle={{
-										height: '300px',
-										textAlign: 'center',
-										paddingTop: '75px',
-										paddingBottom: '75px',
-									}}
-								/>
-							</div>
-						)}
-						{!loadingrecDocs && recDocs.length === 0 && (
-							<div className="col-xs-12" style={{ height: '140px' }}>
-								<Typography style={styles.containerText}>
-									Try favoriting more documents to see your personalized recommendations.
+								>
+									{source.data_source_s}
 								</Typography>
-							</div>
-						)}
-					</GameChangerThumbnailRow>
-					<GameChangerThumbnailRow links={lastOpened} title="Recently Viewed" width="215px">
-						{lastOpened.length > 0 &&
-							lastOpened[0].imgSrc &&
-							lastOpened.map((pub) => (
-								<div className="topPublication">
-									{pub.imgSrc !== 'error' ? (
-										<img className="image" src={pub.imgSrc} alt="thumbnail" title={pub.name} />
-									) : (
-										<div className="image">{pub.name}</div>
-									)}
+							</SourceContainer>
+						))}
+					{crawlerSources.length === 0 && (
+						<div className="col-xs-12" style={{ height: '140px' }}>
+							<LoadingIndicator
+								customColor={gcOrange}
+								inline={true}
+								containerStyle={{ height: '140px', textAlign: 'center' }}
+							/>
+						</div>
+					)}
+				</GameChangerThumbnailRow>
+				<GameChangerThumbnailRow links={recentSearches} title="Recent Searches" width="300px">
+					{recentSearches.map((search) => renderRecentSearches(search, state, dispatch))}
+				</GameChangerThumbnailRow>
+				<GameChangerThumbnailRow links={searchMajorPubs} title="Popular Publications" width="215px">
+					{searchMajorPubs.length > 0 &&
+						searchMajorPubs[0].imgSrc &&
+						searchMajorPubs.map((pub) => (
+							<div className="topPublication">
+								{pub.imgSrc !== 'error' ? (
+									<img className="image" src={pub.imgSrc} alt="thumbnail" title={pub.name} />
+								) : (
+									<div className="image">{pub.name}</div>
+								)}
 
-									<div
-										className="hover-overlay"
-										onClick={() => {
-											trackEvent(
-												getTrackingNameForFactory(cloneData.clone_name),
-												'PublicationOpened',
-												pub.name
-											);
-											pub.imgSrc !== DefaultPub
-												? window.open(
-														`#/gamechanger-details?cloneName=${cloneData.clone_name}&type=document&documentName=${pub.id}`
-												  )
-												: setState(dispatch, { searchText: pub.name, runSearch: true });
-										}}
-									>
-										<div className="hover-text">{formatString(pub.name)}</div>
-									</div>
-								</div>
-							))}
-						{loadingLastOpened && lastOpened.length === 0 && (
-							<div className="col-xs-12">
-								<LoadingIndicator
-									customColor={gcOrange}
-									inline={true}
-									containerStyle={{
-										height: '300px',
-										textAlign: 'center',
-										paddingTop: '75px',
-										paddingBottom: '75px',
-									}}
-								/>
-							</div>
-						)}
-						{!loadingLastOpened && lastOpened.length === 0 && (
-							<div className="col-xs-12" style={{ height: '140px' }}>
-								<Typography style={styles.containerText}>No recent documents to show.</Typography>
-							</div>
-						)}
-					</GameChangerThumbnailRow>
-
-					<GameChangerThumbnailRow links={crawlerSources} title="Sources" width="300px">
-						{crawlerSources.length > 0 &&
-							crawlerSources[0].imgSrc &&
-							crawlerSources.map((source) => (
-								<SourceContainer
+								<div
+									className="hover-overlay"
 									onClick={() => {
 										trackEvent(
 											getTrackingNameForFactory(cloneData.clone_name),
-											'SourceOpened',
-											source.data_source_s
+											'PublicationOpened',
+											pub.name
 										);
+										// window.open(`/#/pdfviewer/gamechanger?filename=${name}&pageNumber=${1}&isClone=${true}&cloneIndex=${cloneData.clone_name}`)
 										window.open(
-											`#/gamechanger-details?cloneName=${
-												cloneData.clone_name
-											}&type=source&sourceName=${source.data_source_s.toLowerCase()}`
+											`#/gamechanger-details?cloneName=${cloneData.clone_name}&type=document&documentName=${pub.id}`
 										);
 									}}
 								>
-									<img src={source.imgSrc} alt={'crawler seal'}></img>
-									<Typography
-										style={{
-											...styles.containerText,
-											color: '#313541',
-											alignSelf: 'center',
-											marginLeft: '20px',
-										}}
-									>
-										{source.data_source_s}
-									</Typography>
-								</SourceContainer>
-							))}
-						{crawlerSources.length === 0 && (
-							<div className="col-xs-12" style={{ height: '140px' }}>
-								<LoadingIndicator
-									customColor={gcOrange}
-									inline={true}
-									containerStyle={{ height: '140px', textAlign: 'center' }}
-								/>
-							</div>
-						)}
-					</GameChangerThumbnailRow>
-					<GameChangerThumbnailRow links={recentSearches} title="Recent Searches" width="300px">
-						{recentSearches.map((search) => renderRecentSearches(search, state, dispatch))}
-					</GameChangerThumbnailRow>
-					<GameChangerThumbnailRow links={searchMajorPubs} title="Popular Publications" width="215px">
-						{searchMajorPubs.length > 0 &&
-							searchMajorPubs[0].imgSrc &&
-							searchMajorPubs.map((pub) => (
-								<div className="topPublication">
-									{pub.imgSrc !== 'error' ? (
-										<img className="image" src={pub.imgSrc} alt="thumbnail" title={pub.name} />
-									) : (
-										<div className="image">{pub.name}</div>
-									)}
-
-									<div
-										className="hover-overlay"
-										onClick={() => {
-											trackEvent(
-												getTrackingNameForFactory(cloneData.clone_name),
-												'PublicationOpened',
-												pub.name
-											);
-											// window.open(`/#/pdfviewer/gamechanger?filename=${name}&pageNumber=${1}&isClone=${true}&cloneIndex=${cloneData.clone_name}`)
-											window.open(
-												`#/gamechanger-details?cloneName=${cloneData.clone_name}&type=document&documentName=${pub.id}`
-											);
-										}}
-									>
-										<div className="hover-text">{formatString(pub.name)}</div>
-									</div>
+									<div className="hover-text">{formatString(pub.name)}</div>
 								</div>
-							))}
-						{searchMajorPubs.length === 0 && (
-							<div className="col-xs-12">
-								<LoadingIndicator
-									customColor={gcOrange}
-									inline={true}
-									containerStyle={{
-										height: '300px',
-										textAlign: 'center',
-										paddingTop: '75px',
-										paddingBottom: '75px',
-									}}
-								/>
+							</div>
+						))}
+					{searchMajorPubs.length === 0 && (
+						<div className="col-xs-12">
+							<LoadingIndicator
+								customColor={gcOrange}
+								inline={true}
+								containerStyle={{
+									height: '300px',
+									textAlign: 'center',
+									paddingTop: '75px',
+									paddingBottom: '75px',
+								}}
+							/>
+						</div>
+					)}
+				</GameChangerThumbnailRow>
+			</div>
+		</div>
+	);
+};
+
+const getViewNames = (props) => {
+	const viewNames = defaultGetViewNames(props);
+	viewNames.push({
+		name: 'Graph',
+		title: 'Graph View',
+		id: 'gcOpenGraphView',
+	});
+	return viewNames;
+};
+
+const getExtraViewPanels = (props) => {
+	const { context } = props;
+	const { state, dispatch } = context;
+	const { cloneData, count, docSearchResults, resultsPage, loading, prevSearchText, searchText } = state;
+	const viewPanels = [];
+	viewPanels.push({
+		panelName: 'Explorer',
+		panel: (
+			<StyledCenterContainer showSideFilters={false}>
+				<div className={'right-container'} style={{ ...styles.tabContainer, margin: '0', height: '800px' }}>
+					<ViewHeader {...props} mainStyles={{ margin: '20px 0 0 0' }} resultsText=" " />
+					<PolicyDocumentExplorer
+						handleSearch={() => setState(dispatch, { runSearch: true })}
+						data={docSearchResults}
+						searchText={searchText}
+						prevSearchText={prevSearchText}
+						totalCount={count}
+						loading={loading}
+						resultsPage={resultsPage}
+						resultsPerPage={RESULTS_PER_PAGE}
+						onPaginationClick={(page) => {
+							setState(dispatch, { resultsPage: page, runSearch: true });
+						}}
+						isClone={true}
+						cloneData={cloneData}
+					/>
+				</div>
+			</StyledCenterContainer>
+		),
+	});
+	viewPanels.push({
+		panelName: 'Graph',
+		panel: (
+			<div key={'graphView'}>
+				{!state.loading && (
+					<StyledCenterContainer showSideFilters={state.showSideFilters}>
+						{state.showSideFilters && (
+							<div className={'left-container'}>
+								<div className={'side-bar-container'}>
+									<div className={'filters-container sidebar-section-title'}>FILTERS</div>
+									<GameChangerSearchMatrix context={context} />
+									{state.sidebarDocTypes.length > 0 && state.sidebarOrgs.length > 0 && (
+										<>
+											<div className={'sidebar-section-title'}>RELATED</div>
+											<GameChangerSideBar context={context} cloneData={state.cloneData} />
+										</>
+									)}
+								</div>
 							</div>
 						)}
-					</GameChangerThumbnailRow>
-				</div>
+						<div className={'right-container'}>
+							<DefaultGraphView context={context} />
+						</div>
+					</StyledCenterContainer>
+				)}
 			</div>
-		);
-	},
+		),
+	});
 
-	getViewNames(props) {
-		const viewNames = defaultMainViewHandler.getViewNames(props);
-		viewNames.push({
-			name: 'Graph',
-			title: 'Graph View',
-			id: 'gcOpenGraphView',
-		});
-		return viewNames;
-	},
+	return viewPanels;
+};
 
-	getExtraViewPanels(props) {
-		const { context } = props;
-		const { state, dispatch } = context;
-		const { cloneData, count, docSearchResults, resultsPage, loading, prevSearchText, searchText } = state;
-		const viewPanels = [];
-		viewPanels.push({
-			panelName: 'Explorer',
-			panel: (
-				<StyledCenterContainer showSideFilters={false}>
-					<div className={'right-container'} style={{ ...styles.tabContainer, margin: '0', height: '800px' }}>
-						<ViewHeader {...props} mainStyles={{ margin: '20px 0 0 0' }} resultsText=" " />
-						<PolicyDocumentExplorer
-							handleSearch={() => setState(dispatch, { runSearch: true })}
-							data={docSearchResults}
-							searchText={searchText}
-							prevSearchText={prevSearchText}
-							totalCount={count}
-							loading={loading}
-							resultsPage={resultsPage}
-							resultsPerPage={RESULTS_PER_PAGE}
-							onPaginationClick={(page) => {
-								setState(dispatch, { resultsPage: page, runSearch: true });
-							}}
-							isClone={true}
-							cloneData={cloneData}
-						/>
-					</div>
-				</StyledCenterContainer>
-			),
-		});
-		viewPanels.push({
-			panelName: 'Graph',
-			panel: (
-				<div key={'graphView'}>
-					{!state.loading && (
-						<StyledCenterContainer showSideFilters={state.showSideFilters}>
-							{state.showSideFilters && (
+const getCardViewPanel = (props) => {
+	const { context } = props;
+	const { state, dispatch } = context;
+	const {
+		activeCategoryTab,
+		cloneData,
+		componentStepNumbers,
+
+		count,
+		docSearchResults,
+		resultsPage,
+		docsLoading,
+		docsPagination,
+
+		entityCount,
+		entitySearchResults,
+		entityPage,
+
+		topicCount,
+		topicSearchResults,
+		topicPage,
+		hideTabs,
+		iframePreviewLink,
+		isCachedResult,
+		loading,
+		selectedCategories,
+		showSideFilters,
+		sidebarOrgs,
+		sidebarDocTypes,
+		timeSinceCache,
+		searchSettings,
+		rawSearchResults,
+	} = state;
+
+	let sideScroll = {
+		height: '72vh',
+	};
+	if (!iframePreviewLink) sideScroll = {};
+	const cacheTip = `Cached result from ${
+		timeSinceCache > 0 ? timeSinceCache + ' hour(s) ago' : 'less than an hour ago'
+	}`;
+
+	return (
+		<div key={'cardView'}>
+			<div key={'cardView'} style={{ marginTop: hideTabs ? 40 : 'auto' }}>
+				<div>
+					<div id="game-changer-content-top" />
+					{!loading && !Boolean(rawSearchResults?.length === 0) && (
+						<StyledCenterContainer showSideFilters={showSideFilters}>
+							{showSideFilters && (
 								<div className={'left-container'}>
 									<div className={'side-bar-container'}>
-										<div className={'filters-container sidebar-section-title'}>FILTERS</div>
 										<GameChangerSearchMatrix context={context} />
-										{state.sidebarDocTypes.length > 0 && state.sidebarOrgs.length > 0 && (
+										{sidebarDocTypes.length > 0 && sidebarOrgs.length > 0 && (
 											<>
 												<div className={'sidebar-section-title'}>RELATED</div>
-												<GameChangerSideBar context={context} cloneData={state.cloneData} />
+												<GameChangerSideBar context={context} cloneData={cloneData} />
 											</>
 										)}
 									</div>
 								</div>
 							)}
 							<div className={'right-container'}>
-								<DefaultGraphView context={context} />
-							</div>
-						</StyledCenterContainer>
-					)}
-				</div>
-			),
-		});
-
-		return viewPanels;
-	},
-
-	getCardViewPanel(props) {
-		const { context } = props;
-		const { state, dispatch } = context;
-		const {
-			activeCategoryTab,
-			cloneData,
-			componentStepNumbers,
-
-			count,
-			docSearchResults,
-			resultsPage,
-			docsLoading,
-			docsPagination,
-
-			entityCount,
-			entitySearchResults,
-			entityPage,
-
-			topicCount,
-			topicSearchResults,
-			topicPage,
-			hideTabs,
-			iframePreviewLink,
-			isCachedResult,
-			loading,
-			selectedCategories,
-			showSideFilters,
-			sidebarOrgs,
-			sidebarDocTypes,
-			timeSinceCache,
-			searchSettings,
-			rawSearchResults,
-		} = state;
-
-		let sideScroll = {
-			height: '72vh',
-		};
-		if (!iframePreviewLink) sideScroll = {};
-		const cacheTip = `Cached result from ${
-			timeSinceCache > 0 ? timeSinceCache + ' hour(s) ago' : 'less than an hour ago'
-		}`;
-
-		return (
-			<div key={'cardView'}>
-				<div key={'cardView'} style={{ marginTop: hideTabs ? 40 : 'auto' }}>
-					<div>
-						<div id="game-changer-content-top" />
-						{!loading && !Boolean(rawSearchResults?.length === 0) && (
-							<StyledCenterContainer showSideFilters={showSideFilters}>
-								{showSideFilters && (
-									<div className={'left-container'}>
-										<div className={'side-bar-container'}>
-											<GameChangerSearchMatrix context={context} />
-											{sidebarDocTypes.length > 0 && sidebarOrgs.length > 0 && (
-												<>
-													<div className={'sidebar-section-title'}>RELATED</div>
-													<GameChangerSideBar context={context} cloneData={cloneData} />
-												</>
-											)}
+								{!hideTabs && <ViewHeader {...props} />}
+								<div
+									className={`row tutorial-step-${componentStepNumbers['Search Results Section']} card-container`}
+									style={{ padding: 0 }}
+								>
+									<div className={'col-xs-12'} style={{ ...sideScroll, padding: 0 }}>
+										<div className="row" style={{ marginLeft: 0, marginRight: 0, padding: 0 }}>
+											{!loading && <GetQAResults context={context} />}
 										</div>
-									</div>
-								)}
-								<div className={'right-container'}>
-									{!hideTabs && <ViewHeader {...props} />}
-									<div
-										className={`row tutorial-step-${componentStepNumbers['Search Results Section']} card-container`}
-										style={{ padding: 0 }}
-									>
-										<div className={'col-xs-12'} style={{ ...sideScroll, padding: 0 }}>
-											<div className="row" style={{ marginLeft: 0, marginRight: 0, padding: 0 }}>
-												{!loading && <GetQAResults context={context} />}
-											</div>
-											{!loading &&
-												(activeCategoryTab === 'Documents' || activeCategoryTab === 'all') &&
-												selectedCategories['Documents'] && (
-													<div
-														className={'col-xs-12'}
-														style={{
-															marginTop: 10,
-															marginLeft: 0,
-															marginRight: 0,
-															paddingRight: 0,
-															paddingLeft: 0,
-														}}
-													>
-														{!searchSettings.isFilterUpdate ? (
-															<SearchSection
-																section={'Documents'}
-																color={'#131E43'}
-																icon={DocumentIcon}
-															>
-																{activeCategoryTab === 'all' ? (
-																	<>
-																		{!docsLoading && !docsPagination ? (
-																			getSearchResults(
-																				docSearchResults,
-																				state,
-																				dispatch
-																			)
-																		) : (
-																			<div className="col-xs-12">
-																				<LoadingIndicator
-																					customColor={gcOrange}
-																				/>
-																			</div>
-																		)}
-																		<div className="gcPagination col-xs-12 text-center">
-																			<Pagination
-																				activePage={resultsPage}
-																				itemsCountPerPage={RESULTS_PER_PAGE}
-																				totalItemsCount={count}
-																				pageRangeDisplayed={8}
-																				onChange={async (page) => {
-																					trackEvent(
-																						getTrackingNameForFactory(
-																							cloneData.clone_name
-																						),
-																						'PaginationChanged',
-																						'page',
-																						page
-																					);
-																					setState(dispatch, {
-																						docsLoading: true,
-																						resultsPage: page,
-																						docsPagination: true,
-																					});
-																				}}
-																			/>
-																		</div>
-																	</>
-																) : (
-																	<>
-																		{getSearchResults(
+										{!loading &&
+											(activeCategoryTab === 'Documents' || activeCategoryTab === 'all') &&
+											selectedCategories['Documents'] && (
+												<div
+													className={'col-xs-12'}
+													style={{
+														marginTop: 10,
+														marginLeft: 0,
+														marginRight: 0,
+														paddingRight: 0,
+														paddingLeft: 0,
+													}}
+												>
+													{!searchSettings.isFilterUpdate ? (
+														<SearchSection
+															section={'Documents'}
+															color={'#131E43'}
+															icon={DocumentIcon}
+														>
+															{activeCategoryTab === 'all' ? (
+																<>
+																	{!docsLoading && !docsPagination ? (
+																		getSearchResults(
 																			docSearchResults,
 																			state,
 																			dispatch
-																		)}
-																		{docsPagination && (
-																			<div className="col-xs-12">
-																				<LoadingIndicator
-																					customColor={gcOrange}
-																					containerStyle={{
-																						margin: '-100px auto',
-																					}}
-																				/>
-																			</div>
-																		)}
-																	</>
-																)}
-															</SearchSection>
-														) : (
-															<div className="col-xs-12">
-																<LoadingIndicator customColor={gcOrange} />
-															</div>
-														)}
-													</div>
-												)}
-
-											{entitySearchResults &&
-												entitySearchResults.length > 0 &&
-												(activeCategoryTab === 'Organizations' ||
-													activeCategoryTab === 'all') &&
-												selectedCategories['Organizations'] && (
-													<div
-														className={'col-xs-12'}
-														style={{
-															marginTop: 10,
-															marginLeft: 0,
-															marginRight: 0,
-															paddingRight: 0,
-															paddingLeft: 0,
-														}}
-													>
-														<SearchSection
-															section={'Organizations'}
-															color={'#376f94'}
-															icon={OrganizationIcon}
-														>
-															{getSearchResults(entitySearchResults, state, dispatch)}
-															<div className="gcPagination col-xs-12 text-center">
-																<Pagination
-																	activePage={entityPage}
-																	itemsCountPerPage={RESULTS_PER_PAGE}
-																	totalItemsCount={entityCount}
-																	pageRangeDisplayed={8}
-																	onChange={async (page) => {
-																		trackEvent(
-																			getTrackingNameForFactory(
-																				cloneData.clone_name
-																			),
-																			'PaginationChanged',
-																			'page',
-																			page
-																		);
-																		setState(dispatch, {
-																			entitiesLoading: true,
-																			entityPage: page,
-																			entityPagination: true,
-																		});
-																	}}
-																/>
-															</div>
+																		)
+																	) : (
+																		<div className="col-xs-12">
+																			<LoadingIndicator customColor={gcOrange} />
+																		</div>
+																	)}
+																	<div className="gcPagination col-xs-12 text-center">
+																		<Pagination
+																			activePage={resultsPage}
+																			itemsCountPerPage={RESULTS_PER_PAGE}
+																			totalItemsCount={count}
+																			pageRangeDisplayed={8}
+																			onChange={async (page) => {
+																				trackEvent(
+																					getTrackingNameForFactory(
+																						cloneData.clone_name
+																					),
+																					'PaginationChanged',
+																					'page',
+																					page
+																				);
+																				setState(dispatch, {
+																					docsLoading: true,
+																					resultsPage: page,
+																					docsPagination: true,
+																				});
+																			}}
+																		/>
+																	</div>
+																</>
+															) : (
+																<>
+																	{getSearchResults(
+																		docSearchResults,
+																		state,
+																		dispatch
+																	)}
+																	{docsPagination && (
+																		<div className="col-xs-12">
+																			<LoadingIndicator
+																				customColor={gcOrange}
+																				containerStyle={{
+																					margin: '-100px auto',
+																				}}
+																			/>
+																		</div>
+																	)}
+																</>
+															)}
 														</SearchSection>
-													</div>
-												)}
+													) : (
+														<div className="col-xs-12">
+															<LoadingIndicator customColor={gcOrange} />
+														</div>
+													)}
+												</div>
+											)}
 
-											{topicSearchResults &&
-												topicSearchResults.length > 0 &&
-												(activeCategoryTab === 'Topics' || activeCategoryTab === 'all') &&
-												selectedCategories['Topics'] && (
-													<div
-														className={'col-xs-12'}
-														style={{
-															marginTop: 10,
-															marginLeft: 0,
-															marginRight: 0,
-															paddingRight: 0,
-															paddingLeft: 0,
-														}}
+										{entitySearchResults &&
+											entitySearchResults.length > 0 &&
+											(activeCategoryTab === 'Organizations' || activeCategoryTab === 'all') &&
+											selectedCategories['Organizations'] && (
+												<div
+													className={'col-xs-12'}
+													style={{
+														marginTop: 10,
+														marginLeft: 0,
+														marginRight: 0,
+														paddingRight: 0,
+														paddingLeft: 0,
+													}}
+												>
+													<SearchSection
+														section={'Organizations'}
+														color={'#376f94'}
+														icon={OrganizationIcon}
 													>
-														<SearchSection
-															section={'Topics'}
-															color={'#4da593'}
-															icon={ApplicationsIcon}
-														>
-															{getSearchResults(topicSearchResults, state, dispatch)}
-															<div className="gcPagination col-xs-12 text-center">
-																<Pagination
-																	activePage={topicPage}
-																	itemsCountPerPage={RESULTS_PER_PAGE}
-																	totalItemsCount={topicCount}
-																	pageRangeDisplayed={8}
-																	onChange={async (page) => {
-																		trackEvent(
-																			getTrackingNameForFactory(
-																				cloneData.clone_name
-																			),
-																			'PaginationChanged',
-																			'page',
-																			page
-																		);
-																		// setState(dispatch, {entitiesLoading: true, entityPage: page, entityPagination: true });
-																	}}
-																/>
-															</div>
-														</SearchSection>
-													</div>
-												)}
-										</div>
+														{getSearchResults(entitySearchResults, state, dispatch)}
+														<div className="gcPagination col-xs-12 text-center">
+															<Pagination
+																activePage={entityPage}
+																itemsCountPerPage={RESULTS_PER_PAGE}
+																totalItemsCount={entityCount}
+																pageRangeDisplayed={8}
+																onChange={async (page) => {
+																	trackEvent(
+																		getTrackingNameForFactory(cloneData.clone_name),
+																		'PaginationChanged',
+																		'page',
+																		page
+																	);
+																	setState(dispatch, {
+																		entitiesLoading: true,
+																		entityPage: page,
+																		entityPagination: true,
+																	});
+																}}
+															/>
+														</div>
+													</SearchSection>
+												</div>
+											)}
+
+										{topicSearchResults &&
+											topicSearchResults.length > 0 &&
+											(activeCategoryTab === 'Topics' || activeCategoryTab === 'all') &&
+											selectedCategories['Topics'] && (
+												<div
+													className={'col-xs-12'}
+													style={{
+														marginTop: 10,
+														marginLeft: 0,
+														marginRight: 0,
+														paddingRight: 0,
+														paddingLeft: 0,
+													}}
+												>
+													<SearchSection
+														section={'Topics'}
+														color={'#4da593'}
+														icon={ApplicationsIcon}
+													>
+														{getSearchResults(topicSearchResults, state, dispatch)}
+														<div className="gcPagination col-xs-12 text-center">
+															<Pagination
+																activePage={topicPage}
+																itemsCountPerPage={RESULTS_PER_PAGE}
+																totalItemsCount={topicCount}
+																pageRangeDisplayed={8}
+																onChange={async (page) => {
+																	trackEvent(
+																		getTrackingNameForFactory(cloneData.clone_name),
+																		'PaginationChanged',
+																		'page',
+																		page
+																	);
+																	// setState(dispatch, {entitiesLoading: true, entityPage: page, entityPagination: true });
+																}}
+															/>
+														</div>
+													</SearchSection>
+												</div>
+											)}
 									</div>
 								</div>
-							</StyledCenterContainer>
-						)}
-						{isCachedResult && (
-							<div style={styles.cachedResultIcon}>
-								<GCTooltip title={cacheTip} placement="right" arrow>
-									<i style={{ cursor: 'pointer' }} className="fa fa-bolt fa-2x" />
-								</GCTooltip>
 							</div>
-						)}
-						{Permissions.isGameChangerAdmin() && !loading && !Boolean(rawSearchResults?.length === 0) && (
-							<div style={styles.cachedResultIcon}>
-								<i
-									style={{ cursor: 'pointer' }}
-									className="fa fa-rocket"
-									onClick={() => setState(dispatch, { showEsQueryDialog: true })}
-								/>
-							</div>
-						)}
-					</div>
+						</StyledCenterContainer>
+					)}
+					{isCachedResult && (
+						<div style={styles.cachedResultIcon}>
+							<GCTooltip title={cacheTip} placement="right" arrow>
+								<i style={{ cursor: 'pointer' }} className="fa fa-bolt fa-2x" />
+							</GCTooltip>
+						</div>
+					)}
+					{Permissions.isGameChangerAdmin() && !loading && !Boolean(rawSearchResults?.length === 0) && (
+						<div style={styles.cachedResultIcon}>
+							<i
+								style={{ cursor: 'pointer' }}
+								className="fa fa-rocket"
+								onClick={() => setState(dispatch, { showEsQueryDialog: true })}
+							/>
+						</div>
+					)}
 				</div>
 			</div>
-		);
-	},
+		</div>
+	);
+};
 
-	getAboutUs(props) {
-		const { state } = props;
-		return <GCAboutUs state={state} />;
-	},
+const getAboutUs = (props) => {
+	const { state } = props;
+	return <GCAboutUs state={state} />;
+};
+
+const getAnalystTools = (context) => {
+	return <AnalystTools context={context} />;
+};
+
+const getDataTracker = (state) => {
+	return <GCDataStatusTracker state={state} />;
+};
+
+const PolicyMainViewHandler = (props) => {
+	const { state, dispatch, cancelToken, setCurrentTime } = props;
+
+	const [pageLoaded, setPageLoaded] = useState(false);
+	const [userProfileHandler, setUserProfileHandler] = useState();
+
+	useEffect(() => {
+		if (state.cloneDataSet && state.historySet && !pageLoaded) {
+			const searchFactory = new SearchHandlerFactory(state.cloneData.search_module);
+			const searchHandler = searchFactory.createHandler();
+
+			const profileFactory = new UserProfileHandlerFactory(state.cloneData.main_view_module);
+
+			const profileHandler = profileFactory.createHandler();
+			setUserProfileHandler(profileHandler);
+
+			handlePageLoad({ state, dispatch, history: state.history, searchHandler, cancelToken });
+			setState(dispatch, { viewNames: getViewNames({ cloneData: state.cloneData }) });
+			setPageLoaded(true);
+		}
+	}, [cancelToken, dispatch, pageLoaded, state]);
+
+	const getViewPanels = () => {
+		const viewPanels = { Card: getCardViewPanel({ context: { state, dispatch } }) };
+
+		const extraViewPanels = getExtraViewPanels({ context: { state, dispatch } });
+		extraViewPanels.forEach(({ panelName, panel }) => {
+			viewPanels[panelName] = panel;
+		});
+
+		return viewPanels;
+	};
+
+	switch (state.pageDisplayed) {
+		case PAGE_DISPLAYED.analystTools:
+			return getNonMainPageOuterContainer(getAnalystTools({ state, dispatch }), state, dispatch);
+		case PAGE_DISPLAYED.dataTracker:
+			return getNonMainPageOuterContainer(getDataTracker(state), state, dispatch);
+		case PAGE_DISPLAYED.userDashboard:
+			return getNonMainPageOuterContainer(
+				userProfileHandler.getUserProfilePage({ state, dispatch }),
+				state,
+				dispatch
+			);
+		case PAGE_DISPLAYED.aboutUs:
+			return getNonMainPageOuterContainer(getAboutUs({ state }), state, dispatch);
+		case PAGE_DISPLAYED.main:
+		default:
+			return getMainView({
+				state,
+				dispatch,
+				setCurrentTime,
+				renderHideTabs,
+				pageLoaded,
+				getViewPanels,
+			});
+	}
 };
 
 export default PolicyMainViewHandler;
